modified files config and main
