+++ config.py
@@ -3,4 +3,4 @@ DEBUG = True
 SQLALCHEMY_DATABASE_URI= 'mysql+pymysql://blogz:MyNewPass@localhost:8889/blogz'
 SQLALCHEMY_ECHO = True
 
-secret_key = 'A0Zr98j/3yX R~XHH!jmN]LWX/,?RU'
+SECRET_KEY = 'A0Zr98j/3yX R~XHH!jmN]LWX/,?RU'
